+++ test/references/init.lua
@@ -129,8 +129,8 @@ end
 
 TEST [[
 local function f()
-    return <?function ()
-    end?>
+    return <~<!function~> ()
+    end!>
 end
 
 local <!f2!> = f()
@@ -138,8 +138,8 @@ local <!f2!> = f()
 
 TEST [[
 local function f()
-    return nil, <?function ()
-    end?>
+    return nil, <~<!function~> ()
+    end!>
 end
 
 local _, <!f2!> = f()